Rename generator and validator module to generators and validators
